+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.2.0] - 2024-12-29
+
+### Added
+
+* Remote DB for data storage
+* Chart for skill counts
+
 ## [0.1.0] - 2024-12-28
 
 ### Added